Add BM25s package support
As alternative to the bm25 implementation in sparse/bm25 a wrapper for the bm25s package is added under sparse/bm25s. The wrapper is built using the native implementation as a template matching most of its methods to methods powered by the bm25s package.
A main difference is that for the bm25s package the tokens are assigned ids by a simple count, whereas fastembed and qdrant expects tokens to be uniquely hashed, so the need to storing the token to id map is avoided. Wrapping to bm25s thus introduced the need to create a hash to token count id map.

Also notice that since bm25s methods also account for the IDF term in the BM25 algos the parallel computation of embeddings might be flawed - this have not yet been tested.